T4.15-4.21

+++ osa4/blogi/controllers/blogs.js
@@ -1,22 +1,35 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 
 })
 
 blogRouter.post('/', async (request, response) => {
+
+  const body = request.body
+
   try {
-    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     if (body.title === undefined || body.url === undefined) {
       return response.status(400).json({ error: 'title or url is missing' })
     }
 
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog({
+      user: user._id,
       title: body.title,
       author: body.author,
       url: body.url,
@@ -24,16 +37,36 @@ blogRouter.post('/', async (request, response) => {
     })
 
     const savedBlog = await blog.save()
-    response.json(savedBlog)
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.json(Blog.format(blog))
   } catch (exception) {
-    console.log(exception)
-    response.status(500).json({ error: 'something went wrong...' })
+    if (exception.name === 'JsonWebTokenError') {
+      response.status(401).json({ error: exception.message })
+    } else {
+      console.log(exception)
+      response.status(500).json({ error: 'something went wrong...' })
+    }
   }
 })
 
 blogRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+    } else {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     response.status(204).end()
   } catch (exception) {
